adding createUser contract call
