+++ frontend/api/index.js
@@ -1,0 +1,40 @@
+import {
+  ApolloClient,
+  InMemoryCache,
+  gql,
+  createHttpLink,
+} from '@apollo/client';
+import { setContext } from '@apollo/client/link/context';
+
+const authLink = setContext((_, { headers }) => {
+  const token = window.localStorage.getItem('lens-auth-token');
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `${token}` : '',
+    },
+  };
+});
+
+const API_URL = 'https://api-mumbai.lens.dev/';
+
+const httpLink = createHttpLink({
+  uri: API_URL,
+});
+
+export const client = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
+
+export * from './lens/authenticate';
+export * from './lens/challenge';
+export * from './lens/createProfile';
+export * from './lens/exploreProfiles';
+export * from './lens/getProfile';
+export * from './lens/getProfileId';
+export * from './lens/getPublications';
+export * from './lens/mirror';
+export * from './lens/refresh';
+
+export * from './ipfonts/createUser';
