+++ frontend/api/lens/authenticate.js
@@ -1,0 +1,8 @@
+export const authenticate = gql`
+    mutation Authenticate($address: EthereumAddress!, $signature: Signature!) {
+        authenticate(request: { address: $address, signature: $signature }) {
+            accessToken
+            refreshToken
+        }
+    }
+`;
