+++ frontend/api/lens/challenge.js
@@ -1,0 +1,7 @@
+export const challenge = gql`
+    query Challenge($address: EthereumAddress!) {
+        challenge(request: { address: $address }) {
+            text
+        }
+    }
+`;
