+++ frontend/api/lens/getProfileId.js
@@ -1,0 +1,9 @@
+export const getProfileId = gql`
+    query Profiles($owner: [EthereumAddress!]) {
+        profiles(request: { ownedBy: $owner }) {
+            items {
+                id
+            }
+        }
+    }
+`;
