+++ frontend/api/lens/refresh.js
@@ -1,0 +1,8 @@
+export const refresh = gql`
+    mutation Refresh($refreshToken: Jwt!) {
+        refresh(request: { refreshToken: $refreshToken }) {
+            accessToken
+            refreshToken
+        }
+    }
+`;
